// packages/common/src/schemas/setting/client-setting.schema.ts
export const clientSettingPath = 'client-setting'
// ...
export const clientSettingSchema = Type.Object(
  {
    id: Type.String({
      format: 'uuid'
    }),
    logo: Type.String(),
    title: Type.String(),
    shortTitle: Type.String(),
    startPath: Type.String(),
    url: Type.String(),
    releaseName: Type.String(),
    siteDescription: Type.String(),
    appleTouchIcon: Type.String(),
    favicon32px: Type.String(),
    favicon16px: Type.String(),
    icon192px: Type.String(),
    icon512px: Type.String(),
    siteManifest: Type.String(),
    safariPinnedTab: Type.String(),
    favicon: Type.String(),
    webmanifestLink: Type.String(),
    swScriptLink: Type.String(),
    appBackground: Type.String(),
    appTitle: Type.String(),
    appSubtitle: Type.String(),
    appDescription: Type.String(),
    appSocialLinks: Type.Array(Type.Ref(clientSocialLinkSchema)),
    gaMeasurementId: Type.String(),
    gtmContainerId: Type.String(),
    gtmAuth: Type.Optional(Type.String({ description: 'For GTM Custom Environments' })),
    gtmPreview: Type.Optional(Type.String({ description: 'For GTM Custom Environments' })),
    themeSettings: Type.Record(Type.String(), Type.Ref(clientThemeOptionsSchema)),
    themeModes: Type.Record(Type.String(), Type.String()),
    key8thWall: Type.String(),
    privacyPolicy: Type.String(),
    homepageLinkButtonEnabled: Type.Boolean(),
    homepageLinkButtonRedirect: Type.String(),
    homepageLinkButtonText: Type.String(),
    createdAt: Type.String({ format: 'date-time' }),
    updatedAt: Type.String({ format: 'date-time' }),
    mediaSettings: Type.Ref(mediaSettingsSchema)
  },
  { $id: 'ClientSetting', additionalProperties: false }
)
export interface ClientSettingType extends Static<typeof clientSettingSchema> {}

export interface ClientSettingDatabaseType
  extends Omit<ClientSettingType, 'appSocialLinks' | 'themeSettings' | 'themeModes'> {
  appSocialLinks: string
  themeSettings: string
  themeModes: string
}